turned existing test apps into samples
