+++ samples/ImportPackage/assets/packages/vector_math
@@ -1,0 +1,1 @@
+/Users/r/.pub-cache/git/vector_math-888f82fd13dc02193f17750b3172f96d8e693110/lib
