+++ samples/ImportPackage/assets/pubspec.yaml
