+++ samples/ImportPackage/include/Resources.h
