+++ samples/ImportPackage/vc2013/ImportPackage.sln
@@ -1,0 +1,21 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.31101.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ImportPackage", "ImportPackage.vcxproj", "{9C892A8E-993C-4997-98C6-D7A9B2585653}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9C892A8E-993C-4997-98C6-D7A9B2585653}.Debug|x64.ActiveCfg = Debug|x64
+		{9C892A8E-993C-4997-98C6-D7A9B2585653}.Debug|x64.Build.0 = Debug|x64
+		{9C892A8E-993C-4997-98C6-D7A9B2585653}.Release|x64.ActiveCfg = Release|x64
+		{9C892A8E-993C-4997-98C6-D7A9B2585653}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
