+++ samples/ImportPackage/xcode/Info.plist
