+++ samples/InfoMap/assets/main.dart
