+++ samples/InfoMap/include/Resources.h
