+++ samples/InfoMap/vc2013/InfoMap.sln
@@ -1,0 +1,21 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.31101.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "InfoMap", "InfoMap.vcxproj", "{06E32240-0B4D-4DAE-BB40-331776774298}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{06E32240-0B4D-4DAE-BB40-331776774298}.Debug|x64.ActiveCfg = Debug|x64
+		{06E32240-0B4D-4DAE-BB40-331776774298}.Debug|x64.Build.0 = Debug|x64
+		{06E32240-0B4D-4DAE-BB40-331776774298}.Release|x64.ActiveCfg = Release|x64
+		{06E32240-0B4D-4DAE-BB40-331776774298}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
