+++ samples/InfoMap/xcode/Info.plist
