+++ samples/common/CinderApp.icns
