+++ samples/common/cinder_app_icon.ico
