--- test/ImportTest/vc2013/Resources.rc
@@ -1,5 +1,0 @@
-#include "../include/Resources.h"
-
-1	ICON	"..\\resources\\cinder_app_icon.ico"
-CIDART_RES_SNAPSHOT_BIN
-CIDART_RES_CINDER_DART
